Modularize walk traversal crate
