+++ crates/walk/src/lib.rs
@@ -83,735 +83,15 @@
 //! - [`rsync_core`](https://docs.rs/rsync-core/latest/rsync_core/) for the
 //!   central orchestration facade.
 
-use std::collections::HashSet;
-use std::error::Error;
-use std::ffi::{OsStr, OsString};
-use std::fmt;
-use std::fs;
-use std::io;
-use std::path::{Path, PathBuf};
-
-/// Configures a filesystem traversal rooted at a specific path.
-#[derive(Clone, Debug)]
-pub struct WalkBuilder {
-    root: PathBuf,
-    follow_symlinks: bool,
-    include_root: bool,
-}
-
-impl WalkBuilder {
-    /// Creates a new builder that will traverse the provided root path.
-    #[must_use]
-    pub fn new<P: Into<PathBuf>>(root: P) -> Self {
-        Self {
-            root: root.into(),
-            follow_symlinks: false,
-            include_root: true,
-        }
-    }
-
-    /// Configures whether directory symlinks should be traversed.
-    ///
-    /// The walker always yields the symlink entry itself. When this option is
-    /// enabled and the symlink points to a directory, the walker also descends
-    /// into the target directory while maintaining the symlink's relative path
-    /// in emitted [`WalkEntry`] values. Canonical paths are tracked to prevent
-    /// infinite loops.
-    #[must_use]
-    pub const fn follow_symlinks(mut self, follow: bool) -> Self {
-        self.follow_symlinks = follow;
-        self
-    }
-
-    /// Controls whether the root entry should be included in the output.
-    ///
-    /// When disabled, traversal starts directly with the root's children.
-    #[must_use]
-    pub const fn include_root(mut self, include: bool) -> Self {
-        self.include_root = include;
-        self
-    }
-
-    /// Builds a [`Walker`] using the configured options.
-    pub fn build(self) -> Result<Walker, WalkError> {
-        let root = absolutize(self.root.clone())?;
-        let metadata = fs::symlink_metadata(&root)
-            .map_err(|error| WalkError::root_metadata(root.clone(), error))?;
-
-        let mut walker = Walker {
-            root,
-            follow_symlinks: self.follow_symlinks,
-            yielded_root: !self.include_root,
-            root_metadata: Some(metadata),
-            stack: Vec::new(),
-            visited: HashSet::new(),
-            finished: false,
-        };
-
-        if let Some(metadata) = walker.root_metadata.as_ref() {
-            let file_type = metadata.file_type();
-            if file_type.is_dir() {
-                walker.push_directory(walker.root.clone(), PathBuf::new(), 0)?;
-            } else if file_type.is_symlink() && walker.follow_symlinks {
-                match fs::metadata(&walker.root) {
-                    Ok(target) if target.is_dir() => {
-                        walker.push_directory(walker.root.clone(), PathBuf::new(), 0)?;
-                    }
-                    Ok(_) => {}
-                    Err(error) => {
-                        return Err(WalkError::metadata(walker.root.clone(), error));
-                    }
-                }
-            }
-        }
-
-        Ok(walker)
-    }
-}
-
-/// Depth-first iterator over filesystem entries.
-pub struct Walker {
-    root: PathBuf,
-    follow_symlinks: bool,
-    yielded_root: bool,
-    root_metadata: Option<fs::Metadata>,
-    stack: Vec<DirectoryState>,
-    visited: HashSet<PathBuf>,
-    finished: bool,
-}
-
-impl Walker {
-    fn push_directory(
-        &mut self,
-        fs_path: PathBuf,
-        relative_prefix: PathBuf,
-        depth: usize,
-    ) -> Result<(), WalkError> {
-        let canonical = fs::canonicalize(&fs_path)
-            .map_err(|error| WalkError::canonicalize(fs_path.clone(), error))?;
-        if !self.visited.insert(canonical) {
-            return Ok(());
-        }
-
-        let state = DirectoryState::new(fs_path, relative_prefix, depth)?;
-        self.stack.push(state);
-        Ok(())
-    }
-
-    fn prepare_entry(
-        &mut self,
-        full_path: PathBuf,
-        relative_path: PathBuf,
-        depth: usize,
-    ) -> Result<WalkEntry, WalkError> {
-        let metadata = fs::symlink_metadata(&full_path)
-            .map_err(|error| WalkError::metadata(full_path.clone(), error))?;
-        let mut next_state = None;
-
-        if metadata.file_type().is_dir() {
-            next_state = Some((full_path.clone(), relative_path.clone(), depth));
-        } else if metadata.file_type().is_symlink() && self.follow_symlinks {
-            match fs::metadata(&full_path) {
-                Ok(target) if target.is_dir() => {
-                    let canonical = fs::canonicalize(&full_path)
-                        .map_err(|error| WalkError::canonicalize(full_path.clone(), error))?;
-                    next_state = Some((canonical, relative_path.clone(), depth));
-                }
-                Ok(_) => {}
-                Err(error) => {
-                    return Err(WalkError::metadata(full_path.clone(), error));
-                }
-            }
-        }
-
-        if let Some((dir_path, rel_prefix, dir_depth)) = next_state {
-            self.push_directory(dir_path, rel_prefix, dir_depth)?;
-        }
-
-        Ok(WalkEntry {
-            full_path,
-            relative_path,
-            metadata,
-            depth,
-            is_root: false,
-        })
-    }
-}
-
-impl Iterator for Walker {
-    type Item = Result<WalkEntry, WalkError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.finished {
-            return None;
-        }
-
-        if !self.yielded_root {
-            self.yielded_root = true;
-            if let Some(metadata) = self.root_metadata.take() {
-                let entry = WalkEntry {
-                    full_path: self.root.clone(),
-                    relative_path: PathBuf::new(),
-                    metadata,
-                    depth: 0,
-                    is_root: true,
-                };
-                return Some(Ok(entry));
-            }
-        }
-
-        loop {
-            let (full_path, relative_path, depth) = {
-                let state = self.stack.last_mut()?;
-
-                if let Some(name) = state.next_name() {
-                    let full_path = state.fs_path.join(&name);
-                    let relative_path = if state.relative_prefix.as_os_str().is_empty() {
-                        PathBuf::from(&name)
-                    } else {
-                        let mut rel = state.relative_prefix.clone();
-                        rel.push(&name);
-                        rel
-                    };
-                    (full_path, relative_path, state.depth + 1)
-                } else {
-                    self.stack.pop();
-                    continue;
-                }
-            };
-
-            match self.prepare_entry(full_path, relative_path, depth) {
-                Ok(entry) => return Some(Ok(entry)),
-                Err(error) => {
-                    self.finished = true;
-                    return Some(Err(error));
-                }
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct DirectoryState {
-    fs_path: PathBuf,
-    relative_prefix: PathBuf,
-    entries: Vec<OsString>,
-    index: usize,
-    depth: usize,
-}
-
-impl DirectoryState {
-    fn new(fs_path: PathBuf, relative_prefix: PathBuf, depth: usize) -> Result<Self, WalkError> {
-        let mut entries = Vec::new();
-        let read_dir =
-            fs::read_dir(&fs_path).map_err(|error| WalkError::read_dir(fs_path.clone(), error))?;
-        for entry in read_dir {
-            let entry = entry.map_err(|error| WalkError::read_dir_entry(fs_path.clone(), error))?;
-            entries.push(entry.file_name());
-        }
-        entries.sort();
-
-        Ok(Self {
-            fs_path,
-            relative_prefix,
-            entries,
-            index: 0,
-            depth,
-        })
-    }
-
-    fn next_name(&mut self) -> Option<OsString> {
-        if let Some(name) = self.entries.get(self.index) {
-            self.index += 1;
-            Some(name.clone())
-        } else {
-            None
-        }
-    }
-}
-
-/// Result of a filesystem traversal step.
-#[derive(Debug)]
-pub struct WalkEntry {
-    full_path: PathBuf,
-    relative_path: PathBuf,
-    metadata: fs::Metadata,
-    depth: usize,
-    is_root: bool,
-}
-
-impl WalkEntry {
-    /// Returns the absolute path to the filesystem entry.
-    #[must_use]
-    pub fn full_path(&self) -> &Path {
-        &self.full_path
-    }
-
-    /// Returns the path relative to the traversal root.
-    #[must_use]
-    pub fn relative_path(&self) -> &Path {
-        &self.relative_path
-    }
-
-    /// Provides access to the [`fs::Metadata`] captured for the entry.
-    #[must_use]
-    pub fn metadata(&self) -> &fs::Metadata {
-        &self.metadata
-    }
-
-    /// Returns the file name associated with the entry, if any.
-    ///
-    /// The root entry of a traversal has no parent directory and therefore
-    /// yields `None`. All other entries return the final component of the
-    /// relative path, matching upstream rsync's expectations when constructing
-    /// file-list nodes.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rsync_walk::WalkBuilder;
-    /// # fn demo() -> Result<(), Box<dyn std::error::Error>> {
-    /// let temp = tempfile::tempdir()?;
-    /// let root = temp.path().join("root");
-    /// std::fs::create_dir(&root)?;
-    /// let mut walker = WalkBuilder::new(&root).build()?;
-    /// let entry = walker.next().unwrap()?;
-    /// assert!(entry.metadata().is_dir());
-    /// assert!(entry.file_name().is_none());
-    /// # Ok(())
-    /// # }
-    /// # demo().unwrap();
-    /// ```
-    #[must_use]
-    pub fn file_name(&self) -> Option<&OsStr> {
-        if self.is_root {
-            None
-        } else {
-            self.relative_path.file_name()
-        }
-    }
-
-    /// Reports the depth of the entry relative to the root (root depth is `0`).
-    #[must_use]
-    pub const fn depth(&self) -> usize {
-        self.depth
-    }
-
-    /// Indicates whether this entry corresponds to the traversal root.
-    #[must_use]
-    pub const fn is_root(&self) -> bool {
-        self.is_root
-    }
-}
-
-/// Error returned when traversal fails.
-#[derive(Debug)]
-pub struct WalkError {
-    kind: WalkErrorKind,
-}
-
-impl WalkError {
-    fn new(kind: WalkErrorKind) -> Self {
-        Self { kind }
-    }
-
-    fn root_metadata(path: PathBuf, source: io::Error) -> Self {
-        Self::new(WalkErrorKind::RootMetadata { path, source })
-    }
-
-    fn read_dir(path: PathBuf, source: io::Error) -> Self {
-        Self::new(WalkErrorKind::ReadDir { path, source })
-    }
-
-    fn read_dir_entry(path: PathBuf, source: io::Error) -> Self {
-        Self::new(WalkErrorKind::ReadDirEntry { path, source })
-    }
-
-    fn metadata(path: PathBuf, source: io::Error) -> Self {
-        Self::new(WalkErrorKind::Metadata { path, source })
-    }
-
-    fn canonicalize(path: PathBuf, source: io::Error) -> Self {
-        Self::new(WalkErrorKind::Canonicalize { path, source })
-    }
-
-    /// Returns the specific failure that terminated traversal.
-    #[must_use]
-    pub fn kind(&self) -> &WalkErrorKind {
-        &self.kind
-    }
-
-    /// Returns the filesystem path associated with the error.
-    ///
-    /// The helper mirrors upstream diagnostics, which always include the
-    /// offending path in walker failures. Callers can forward the returned path
-    /// directly into higher-level error messages without having to pattern match
-    /// on [`WalkErrorKind`].
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use rsync_walk::WalkBuilder;
-    ///
-    /// let result = WalkBuilder::new("./definitely_missing_root").build();
-    /// let error = match result {
-    ///     Ok(_) => panic!("missing root yields error"),
-    ///     Err(error) => error,
-    /// };
-    /// assert!(error.path().ends_with("definitely_missing_root"));
-    /// ```
-    #[must_use]
-    pub fn path(&self) -> &Path {
-        self.kind.path()
-    }
-}
-
-impl fmt::Display for WalkError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self.kind {
-            WalkErrorKind::RootMetadata { path, source } => {
-                write!(
-                    f,
-                    "failed to inspect traversal root '{}': {}",
-                    path.display(),
-                    source
-                )
-            }
-            WalkErrorKind::ReadDir { path, source } => {
-                write!(
-                    f,
-                    "failed to read directory '{}': {}",
-                    path.display(),
-                    source
-                )
-            }
-            WalkErrorKind::ReadDirEntry { path, source } => {
-                write!(
-                    f,
-                    "failed to read entry in '{}': {}",
-                    path.display(),
-                    source
-                )
-            }
-            WalkErrorKind::Metadata { path, source } => {
-                write!(
-                    f,
-                    "failed to inspect metadata for '{}': {}",
-                    path.display(),
-                    source
-                )
-            }
-            WalkErrorKind::Canonicalize { path, source } => {
-                write!(f, "failed to canonicalize '{}': {}", path.display(), source)
-            }
-        }
-    }
-}
-
-impl Error for WalkError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match &self.kind {
-            WalkErrorKind::RootMetadata { source, .. }
-            | WalkErrorKind::ReadDir { source, .. }
-            | WalkErrorKind::ReadDirEntry { source, .. }
-            | WalkErrorKind::Metadata { source, .. }
-            | WalkErrorKind::Canonicalize { source, .. } => Some(source),
-        }
-    }
-}
-
-/// Classification of traversal failures.
-#[derive(Debug)]
-pub enum WalkErrorKind {
-    /// Failed to query metadata for the traversal root.
-    RootMetadata {
-        /// Path that failed to provide metadata.
-        path: PathBuf,
-        /// Underlying error emitted by the operating system.
-        source: io::Error,
-    },
-    /// Failed to read the contents of a directory.
-    ReadDir {
-        /// Directory whose contents could not be read.
-        path: PathBuf,
-        /// Underlying error emitted by the operating system.
-        source: io::Error,
-    },
-    /// Failed to obtain a directory entry during iteration.
-    ReadDirEntry {
-        /// Directory containing the problematic entry.
-        path: PathBuf,
-        /// Underlying error emitted by the operating system.
-        source: io::Error,
-    },
-    /// Failed to retrieve metadata for an entry.
-    Metadata {
-        /// Path whose metadata could not be retrieved.
-        path: PathBuf,
-        /// Underlying error emitted by the operating system.
-        source: io::Error,
-    },
-    /// Failed to canonicalize a directory path while preventing cycles.
-    Canonicalize {
-        /// Directory path that failed to canonicalize.
-        path: PathBuf,
-        /// Underlying error emitted by the operating system.
-        source: io::Error,
-    },
-}
-
-impl WalkErrorKind {
-    /// Returns the filesystem path tied to the failure.
-    #[must_use]
-    pub fn path(&self) -> &Path {
-        match self {
-            WalkErrorKind::RootMetadata { path, .. }
-            | WalkErrorKind::ReadDir { path, .. }
-            | WalkErrorKind::ReadDirEntry { path, .. }
-            | WalkErrorKind::Metadata { path, .. }
-            | WalkErrorKind::Canonicalize { path, .. } => path,
-        }
-    }
-}
-
-fn absolutize(path: PathBuf) -> Result<PathBuf, WalkError> {
-    if path.is_absolute() {
-        Ok(path)
-    } else {
-        let cwd = std::env::current_dir()
-            .map_err(|error| WalkError::canonicalize(PathBuf::from("."), error))?;
-        Ok(cwd.join(path))
-    }
-}
+mod builder;
+mod entry;
+mod error;
+mod walker;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-    use std::path::Path;
+mod tests;
 
-    fn collect_relative_paths(walker: Walker) -> Vec<PathBuf> {
-        let mut paths = Vec::new();
-        for entry in walker {
-            let entry = entry.expect("walker entry");
-            if entry.is_root() {
-                continue;
-            }
-            paths.push(entry.relative_path().to_path_buf());
-        }
-        paths
-    }
-
-    #[test]
-    fn walk_errors_when_root_missing() {
-        let builder = WalkBuilder::new("/nonexistent/path/for/walker");
-        let error = match builder.build() {
-            Ok(_) => panic!("missing root should fail"),
-            Err(error) => error,
-        };
-        assert!(matches!(error.kind(), WalkErrorKind::RootMetadata { .. }));
-        assert_eq!(error.path(), Path::new("/nonexistent/path/for/walker"));
-        assert_eq!(
-            error.kind().path(),
-            Path::new("/nonexistent/path/for/walker")
-        );
-    }
-
-    #[test]
-    fn walk_single_file_emits_root_entry() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let file = temp.path().join("file.txt");
-        fs::write(&file, b"contents").expect("write");
-
-        let mut walker = WalkBuilder::new(&file).build().expect("build walker");
-        let entry = walker.next().expect("entry").expect("entry ok");
-        assert!(entry.is_root());
-        assert!(entry.relative_path().as_os_str().is_empty());
-        assert_eq!(entry.full_path(), file);
-        assert!(walker.next().is_none());
-    }
-
-    #[test]
-    fn walk_directory_yields_deterministic_order() {
-        let temp = tempfile::tempdir().expect("tempdir");
-        let root = temp.path().join("root");
-        fs::create_dir(&root).expect("create root");
-        let dir_a = root.join("a");
-        let dir_b = root.join("b");
-        let file_c = root.join("c.txt");
-        fs::create_dir(&dir_a).expect("dir a");
-        fs::create_dir(&dir_b).expect("dir b");
-        fs::write(dir_a.join("inner.txt"), b"data").expect("write inner");
-        fs::write(&file_c, b"data").expect("write file");
-
-        let walker = WalkBuilder::new(&root).build().expect("build walker");
-        let paths = collect_relative_paths(walker);
-        assert_eq!(
-            paths,
-            vec![
-                PathBuf::from("a"),
-                PathBuf::from("a/inner.txt"),
-                PathBuf::from("b"),
-                PathBuf::from("c.txt"),
-            ]
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn walk_does_not_follow_symlink_by_default() {
-        use std::os::unix::fs::symlink;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let root = temp.path().join("root");
-        let target = temp.path().join("target");
-        fs::create_dir(&root).expect("create root");
-        fs::create_dir(&target).expect("create target");
-        fs::write(target.join("inner.txt"), b"data").expect("write inner");
-        symlink(&target, root.join("link")).expect("create symlink");
-
-        let walker = WalkBuilder::new(&root).build().expect("build walker");
-        let paths = collect_relative_paths(walker);
-        assert_eq!(paths, vec![PathBuf::from("link")]);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn walk_follows_symlink_when_enabled() {
-        use std::os::unix::fs::symlink;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let root = temp.path().join("root");
-        let target = temp.path().join("target");
-        fs::create_dir(&root).expect("create root");
-        fs::create_dir(&target).expect("create target");
-        fs::write(target.join("inner.txt"), b"data").expect("write inner");
-        symlink(&target, root.join("link")).expect("create symlink");
-
-        let walker = WalkBuilder::new(&root)
-            .follow_symlinks(true)
-            .build()
-            .expect("build walker");
-        let paths = collect_relative_paths(walker);
-        assert_eq!(
-            paths,
-            vec![PathBuf::from("link"), PathBuf::from("link/inner.txt")]
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn walk_root_symlink_followed_when_enabled() {
-        use std::os::unix::fs::symlink;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let target = temp.path().join("target");
-        fs::create_dir(&target).expect("create target");
-        fs::write(target.join("file.txt"), b"data").expect("write file");
-
-        let link = temp.path().join("link");
-        symlink(&target, &link).expect("create symlink");
-
-        let walker = WalkBuilder::new(&link)
-            .follow_symlinks(true)
-            .build()
-            .expect("build walker");
-        let paths = collect_relative_paths(walker);
-
-        assert_eq!(paths, vec![PathBuf::from("file.txt")]);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn walk_root_symlink_preserves_full_paths() {
-        use std::os::unix::fs::symlink;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let target = temp.path().join("target");
-        fs::create_dir(&target).expect("create target");
-        let file = target.join("file.txt");
-        fs::write(&file, b"data").expect("write file");
-
-        let link = temp.path().join("link");
-        symlink(&target, &link).expect("create symlink");
-
-        let mut walker = WalkBuilder::new(&link)
-            .follow_symlinks(true)
-            .build()
-            .expect("build walker");
-
-        let root = walker.next().expect("root entry").expect("root ok");
-        assert!(root.is_root());
-        assert!(root.metadata().file_type().is_symlink());
-        assert_eq!(root.full_path(), link.as_path());
-        assert!(root.relative_path().as_os_str().is_empty());
-
-        let child = walker.next().expect("child entry").expect("child ok");
-        assert_eq!(child.relative_path(), std::path::Path::new("file.txt"));
-        assert_eq!(child.full_path(), link.join("file.txt"));
-        assert!(child.metadata().is_file());
-
-        assert!(walker.next().is_none());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn walk_root_symlink_not_followed_by_default() {
-        use std::os::unix::fs::symlink;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let target = temp.path().join("target");
-        fs::create_dir(&target).expect("create target");
-        fs::write(target.join("file.txt"), b"data").expect("write file");
-
-        let link = temp.path().join("link");
-        symlink(&target, &link).expect("create symlink");
-
-        let mut walker = WalkBuilder::new(&link).build().expect("build walker");
-        let root = walker.next().expect("root entry").expect("root ok");
-        assert!(root.is_root());
-        assert!(root.metadata().file_type().is_symlink());
-        assert!(walker.next().is_none());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn walk_detects_symlink_cycles() {
-        use std::os::unix::fs::symlink;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let root = temp.path().join("root");
-        fs::create_dir(&root).expect("create root");
-        let _ = symlink(&root, root.join("self"));
-
-        let walker = WalkBuilder::new(&root)
-            .follow_symlinks(true)
-            .build()
-            .expect("build walker");
-        let paths = collect_relative_paths(walker);
-        assert_eq!(paths, vec![PathBuf::from("self")]);
-    }
-
-    #[test]
-    fn walk_entry_file_name_matches_tail_component() {
-        use std::ffi::OsStr;
-
-        let temp = tempfile::tempdir().expect("tempdir");
-        let root = temp.path().join("root");
-        let nested_dir = root.join("nested");
-        let nested_file = nested_dir.join("file.txt");
-        fs::create_dir_all(&nested_dir).expect("create nested");
-        fs::write(&nested_file, b"data").expect("write nested file");
-
-        let mut walker = WalkBuilder::new(&root).build().expect("build walker");
-        let root_entry = walker.next().expect("root entry").expect("root ok");
-        assert!(root_entry.is_root());
-        assert!(root_entry.file_name().is_none());
-
-        let dir_entry = walker.next().expect("dir entry").expect("dir ok");
-        assert_eq!(dir_entry.file_name(), Some(OsStr::new("nested")));
-
-        let file_entry = walker.next().expect("file entry").expect("file ok");
-        assert_eq!(file_entry.file_name(), Some(OsStr::new("file.txt")));
-    }
-}
+pub use crate::builder::WalkBuilder;
+pub use crate::entry::WalkEntry;
+pub use crate::error::{WalkError, WalkErrorKind};
+pub use crate::walker::Walker;
